refactor(web-ui): replace include!(\"main.rs\") with idiomatic lib + thin shim
The web-ui crate previously bridged its library surface (consumed by
`assistant-cli`, integration tests, and `crates/web-ui/tests/`) and its
binary entrypoint by having `lib.rs` consist of:

    #![allow(dead_code)]
    include!(\"main.rs\");

This hack required a crate-wide `dead_code` allow (silencing real
warnings) and gave the binary and library two co-existing copies of
every symbol through textual inclusion. The pattern was on the
codebase-analysis punch list as one of the top hygiene smells.

Restructure to the idiomatic shape:

- `src/lib.rs` now holds all the real logic: module declarations
  (`api`, `auth`, `oauth`, `install`, `sw_version`, …), shared types
  (`AppState`, `OrchestratorTurnClient`, …), and the public entry
  points `run_from_env` / `run_from_iter` already consumed by the CLI
  and integration tests.
- `src/main.rs` is a 4-line shim that calls
  `assistant_web_ui::run_from_env().await`.
- Drop the crate-wide `#![allow(dead_code)]`.

Four genuine dead-code warnings surfaced after removing the blanket
allow. Each is annotated with a per-field `#[allow(dead_code)]` and a
comment explaining the spec/forward-compat reason it's kept:

- `AppState::trace_limit`, `AppState::log_limit` — start-up config
  threaded through state but not yet read by any route.
- `GetTaskQuery::history_length`, `GetTaskQuery::tenant` — A2A
  protocol fields accepted for forward-compat, OpenAPI surface match.
- `AuthorizeQuery::code_challenge_method` — RFC 7636 §4.3 field; we
  assume `S256` implicitly.
- `RevokeRequest::token_type_hint`, `RevokeRequest::client_id` —
  RFC 7009 §2.1 fields; revocation matches by token value alone.

Tests:
- `cargo test -p assistant-web-ui --lib` — 241 pass
- `cargo test -p assistant-web-ui --test sw_version_injection` — 4 pass
  (this is the external test that motivated the original hack)
